v2.2 1、完成菜单、按钮的按照权限进行生成

+++ src/main/java/com/bangnd/util/repository/BusinessTypeReository.java
@@ -3,5 +3,5 @@ package com.bangnd.util.repository;
 import com.bangnd.util.entity.BusinessType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessTypeReository extends JpaRepository<BusinessType, Long> {
+public interface BusinessTypeReository extends JpaRepository<BusinessType, Integer> {
 }
